requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
openpyxl>=3.1.2
python-dateutil>=2.9.0
pytest>=8.0.0
xlsxwriter>=3.2.0
